+++ backend/canisters/airdrop_bot/impl/src/jobs/execute_airdrop.rs
@@ -67,7 +67,6 @@ async fn prepare_airdrop(config: AirdropConfig, user_index_canister_id: Canister
             avatar: OptionUpdate::NoChange,
             permissions_v2: None,
             events_ttl: OptionUpdate::NoChange,
-            gate: OptionUpdate::SetToSome(AccessGate::Locked),
             gate_config: OptionUpdate::SetToSome(AccessGate::Locked.into()),
             public: None,
             messages_visible_to_non_members: None,