+++ backend/canisters/community/impl/src/updates/update_channel.rs
@@ -46,9 +46,7 @@ fn update_channel_impl(mut args: Args, state: &mut RuntimeState) -> Response {
 
         if let Some(member) = state.data.members.get(caller) {
             let now = state.env.now();
-            let gate_config_updates =
-                if args.gate_config.has_update() { args.gate_config } else { args.gate.map(|g| g.into()) };
-            let has_gate_config_updates = gate_config_updates.has_update();
+            let has_gate_config_updates = args.gate_config.has_update();
 
             let prev_gate_config = channel.chat.gate_config.value.clone();
 
@@ -59,7 +57,7 @@ fn update_channel_impl(mut args: Args, state: &mut RuntimeState) -> Response {
                 args.rules,
                 args.avatar,
                 args.permissions_v2,
-                gate_config_updates.map(|gc| gc.into()),
+                args.gate_config.map(|gc| gc.into()),
                 args.public,
                 args.messages_visible_to_non_members,
                 args.events_ttl,