+++ backend/canisters/group_index/impl/src/updates/c2c_create_group.rs
@@ -36,7 +36,6 @@ async fn c2c_create_group(args: Args) -> Response {
         messages_visible_to_non_members: args.messages_visible_to_non_members,
         permissions_v2: args.permissions_v2,
         events_ttl: args.events_ttl,
-        gate: args.gate.clone(),
         gate_config: args.gate_config.clone(),
     };